+++ module/Core/test-db/Repository/ShortUrlRepositoryTest.php
@@ -14,6 +14,7 @@ use Shlinkio\Shlink\Core\Entity\Visit;
 use Shlinkio\Shlink\Core\Model\ShortUrlIdentifier;
 use Shlinkio\Shlink\Core\Model\ShortUrlMeta;
 use Shlinkio\Shlink\Core\Model\ShortUrlsOrdering;
+use Shlinkio\Shlink\Core\Model\ShortUrlsParams;
 use Shlinkio\Shlink\Core\Model\Visitor;
 use Shlinkio\Shlink\Core\Repository\ShortUrlRepository;
 use Shlinkio\Shlink\Core\ShortUrl\Resolver\PersistenceShortUrlRelationResolver;
@@ -127,22 +128,30 @@ class ShortUrlRepositoryTest extends DatabaseTestCase
 
         self::assertCount(1, $this->repo->findList(2, 2));
 
-        $result = $this->repo->findList(null, null, null, [], ShortUrlsOrdering::fromRawData([
+        $result = $this->repo->findList(null, null, null, [], null, ShortUrlsOrdering::fromRawData([
             'orderBy' => 'visits-DESC',
         ]));
         self::assertCount(3, $result);
         self::assertSame($bar, $result[0]);
 
-        $result = $this->repo->findList(null, null, null, [], null, DateRange::withEndDate(Chronos::now()->subDays(2)));
+        $result = $this->repo->findList(null, null, null, [], null, null, DateRange::withEndDate(
+            Chronos::now()->subDays(2),
+        ));
         self::assertCount(1, $result);
-        self::assertEquals(1, $this->repo->countList(null, [], DateRange::withEndDate(Chronos::now()->subDays(2))));
+        self::assertEquals(1, $this->repo->countList(null, [], null, DateRange::withEndDate(
+            Chronos::now()->subDays(2),
+        )));
         self::assertSame($foo2, $result[0]);
 
         self::assertCount(
             2,
-            $this->repo->findList(null, null, null, [], null, DateRange::withStartDate(Chronos::now()->subDays(2))),
+            $this->repo->findList(null, null, null, [], null, null, DateRange::withStartDate(
+                Chronos::now()->subDays(2),
+            )),
         );
-        self::assertEquals(2, $this->repo->countList(null, [], DateRange::withStartDate(Chronos::now()->subDays(2))));
+        self::assertEquals(2, $this->repo->countList(null, [], null, DateRange::withStartDate(
+            Chronos::now()->subDays(2),
+        )));
     }
 
     /** @test */
@@ -155,7 +164,7 @@ class ShortUrlRepositoryTest extends DatabaseTestCase
 
         $this->getEntityManager()->flush();
 
-        $result = $this->repo->findList(null, null, null, [], ShortUrlsOrdering::fromRawData([
+        $result = $this->repo->findList(null, null, null, [], null, ShortUrlsOrdering::fromRawData([
             'orderBy' => 'longUrl-ASC',
         ]));
 
@@ -164,6 +173,71 @@ class ShortUrlRepositoryTest extends DatabaseTestCase
         self::assertEquals('b', $result[1]->getLongUrl());
         self::assertEquals('c', $result[2]->getLongUrl());
         self::assertEquals('z', $result[3]->getLongUrl());
+    }
+
+    /** @test */
+    public function findListReturnsOnlyThoseWithMatchingTags(): void
+    {
+        $shortUrl1 = ShortUrl::fromMeta(ShortUrlMeta::fromRawData([
+            'longUrl' => 'foo1',
+            'tags' => ['foo', 'bar'],
+        ]), $this->relationResolver);
+        $this->getEntityManager()->persist($shortUrl1);
+        $shortUrl2 = ShortUrl::fromMeta(ShortUrlMeta::fromRawData([
+            'longUrl' => 'foo2',
+            'tags' => ['foo', 'baz'],
+        ]), $this->relationResolver);
+        $this->getEntityManager()->persist($shortUrl2);
+        $shortUrl3 = ShortUrl::fromMeta(ShortUrlMeta::fromRawData([
+            'longUrl' => 'foo3',
+            'tags' => ['foo'],
+        ]), $this->relationResolver);
+        $this->getEntityManager()->persist($shortUrl3);
+        $shortUrl4 = ShortUrl::fromMeta(ShortUrlMeta::fromRawData([
+            'longUrl' => 'foo4',
+            'tags' => ['bar', 'baz'],
+        ]), $this->relationResolver);
+        $this->getEntityManager()->persist($shortUrl4);
+        $shortUrl5 = ShortUrl::fromMeta(ShortUrlMeta::fromRawData([
+            'longUrl' => 'foo5',
+            'tags' => ['bar', 'baz'],
+        ]), $this->relationResolver);
+        $this->getEntityManager()->persist($shortUrl5);
+
+        $this->getEntityManager()->flush();
+
+        self::assertCount(5, $this->repo->findList(null, null, null, ['foo', 'bar']));
+        self::assertCount(5, $this->repo->findList(null, null, null, ['foo', 'bar'], ShortUrlsParams::TAGS_MODE_ANY));
+        self::assertCount(1, $this->repo->findList(null, null, null, ['foo', 'bar'], ShortUrlsParams::TAGS_MODE_ALL));
+        self::assertEquals(5, $this->repo->countList(null, ['foo', 'bar']));
+        self::assertEquals(5, $this->repo->countList(null, ['foo', 'bar'], ShortUrlsParams::TAGS_MODE_ANY));
+        self::assertEquals(1, $this->repo->countList(null, ['foo', 'bar'], ShortUrlsParams::TAGS_MODE_ALL));
+
+        self::assertCount(4, $this->repo->findList(null, null, null, ['bar', 'baz']));
+        self::assertCount(4, $this->repo->findList(null, null, null, ['bar', 'baz'], ShortUrlsParams::TAGS_MODE_ANY));
+        self::assertCount(2, $this->repo->findList(null, null, null, ['bar', 'baz'], ShortUrlsParams::TAGS_MODE_ALL));
+        self::assertEquals(4, $this->repo->countList(null, ['bar', 'baz']));
+        self::assertEquals(4, $this->repo->countList(null, ['bar', 'baz'], ShortUrlsParams::TAGS_MODE_ANY));
+        self::assertEquals(2, $this->repo->countList(null, ['bar', 'baz'], ShortUrlsParams::TAGS_MODE_ALL));
+
+        self::assertCount(5, $this->repo->findList(null, null, null, ['foo', 'bar', 'baz']));
+        self::assertCount(5, $this->repo->findList(
+            null,
+            null,
+            null,
+            ['foo', 'bar', 'baz'],
+            ShortUrlsParams::TAGS_MODE_ANY,
+        ));
+        self::assertCount(0, $this->repo->findList(
+            null,
+            null,
+            null,
+            ['foo', 'bar', 'baz'],
+            ShortUrlsParams::TAGS_MODE_ALL,
+        ));
+        self::assertEquals(5, $this->repo->countList(null, ['foo', 'bar', 'baz']));
+        self::assertEquals(5, $this->repo->countList(null, ['foo', 'bar', 'baz'], ShortUrlsParams::TAGS_MODE_ANY));
+        self::assertEquals(0, $this->repo->countList(null, ['foo', 'bar', 'baz'], ShortUrlsParams::TAGS_MODE_ALL));
     }
 
     /** @test */